Refactor imports and update model references for accounts and movies
