+++ src/database/__init__.py
@@ -1,35 +1,18 @@
 import os
 
 from database.models.base import Base
-from database.models.accounts import (
-    UserModel,
-    UserGroupModel,
-    UserGroupEnum,
-    ActivationTokenModel,
-    PasswordResetTokenModel,
-    RefreshTokenModel,
-    UserProfileModel
-)
-from database.models.movies import (
-    MovieModel,
-    GenreModel,
-)
 from database.session_sqlite import reset_sqlite_database as reset_database
 from database.validators import accounts as accounts_validators
 
 environment = os.getenv("ENVIRONMENT", "developing")
 
 if environment == "testing":
-    from database.session_sqlite import (
-        get_sqlite_db_contextmanager as get_db_contextmanager,
-        get_sqlite_db as get_db,
-        get_sync_sqlite_db as get_sync_db,
-        get_sync_sqlite_db_contextmanager as get_sync_db_contextmanager
-    )
+    from database.session_sqlite import get_sqlite_db as get_db
+    from database.session_sqlite import get_sqlite_db_contextmanager as get_db_contextmanager
+    from database.session_sqlite import get_sync_sqlite_db as get_sync_db
+    from database.session_sqlite import get_sync_sqlite_db_contextmanager as get_sync_db_contextmanager
 else:
-    from database.session_postgresql import (
-        get_postgresql_db_contextmanager as get_db_contextmanager,
-        get_postgresql_db as get_db,
-        get_sync_postgresql_db as get_sync_db,
-        get_sync_postgresql_db_contextmanager as get_sync_db_contextmanager
-    )
+    from database.session_postgresql import get_postgresql_db as get_db
+    from database.session_postgresql import get_postgresql_db_contextmanager as get_db_contextmanager
+    from database.session_postgresql import get_sync_postgresql_db as get_sync_db
+    from database.session_postgresql import get_sync_postgresql_db_contextmanager as get_sync_db_contextmanager
