+++ src/pagination/pages.py
@@ -7,7 +7,7 @@ from fastapi_pagination.bases import AbstractPage, AbstractParams, RawParams
 from fastapi_pagination.customization import CustomizedPage, UseFieldsAliases
 from pydantic import BaseModel, Field
 
-from schemas import MovieDetailSchema
+from schemas.movies import MovieDetailSchema
 
 T = TypeVar("T")
