--- src/routes/__init__.py
@@ -1,3 +1,0 @@
-from routes.accounts import router as accounts_router
-from routes.movies import router as movie_router
-from routes.profiles import router as profiles_router
