+++ src/scheduler/tasks.py
@@ -4,7 +4,8 @@ from datetime import datetime, timezone
 from sqlalchemy import delete
 from sqlalchemy.exc import SQLAlchemyError
 
-from database import ActivationTokenModel, get_sync_db_contextmanager
+from database import get_sync_db_contextmanager
+from database.models.accounts import ActivationTokenModel
 from scheduler.celery_app import celery_app
 
 logger = logging.getLogger(__name__)
