--- src/schemas/__init__.py
@@ -1,20 +1,0 @@
-from schemas.accounts import (
-    MessageResponseSchema,
-    PasswordResetCompleteRequestSchema,
-    PasswordResetRequestSchema,
-    ResendActivationRequestSchema,
-    TokenRefreshRequestSchema,
-    TokenRefreshResponseSchema,
-    UserActivationRequestSchema,
-    UserLoginRequestSchema,
-    UserLoginResponseSchema,
-    UserRegistrationRequestSchema,
-    UserRegistrationResponseSchema,
-)
-from schemas.movies import (
-    MovieCreateSchema,
-    MovieDetailSchema,
-    MovieListItemSchema,
-    MovieListResponseSchema,
-    MovieUpdateSchema,
-)
