+++ src/schemas/movies.py
@@ -36,8 +36,7 @@ class GenreReadSchema(GenreBaseSchema):
     id: int
     movie_count: Optional[int] = Field(None, examples=[12])
 
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class StarBaseSchema(BaseModel):
@@ -61,8 +60,7 @@ class StarDeleteSchema(StarBaseSchema):
 class StarReadSchema(StarBaseSchema):
     id: int
 
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class DirectorBaseSchema(BaseModel):
@@ -86,8 +84,7 @@ class DirectorDeleteSchema(DirectorBaseSchema):
 class DirectorReadSchema(DirectorBaseSchema):
     id: int
 
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class CertificationBaseSchema(BaseModel):
@@ -111,8 +108,7 @@ class CertificationDeleteSchema(CertificationBaseSchema):
 class CertificationReadSchema(CertificationBaseSchema):
     id: int
 
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class MovieBaseSchema(BaseModel):
