+++ src/schemas/payments.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 from enum import Enum
 from typing import Optional
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
 
 class PaymentItemBaseSchema(BaseModel):
@@ -11,8 +11,7 @@ class PaymentItemBaseSchema(BaseModel):
     order_item_id: int
     price_at_payment: Decimal
 
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class PaymentStatusSchema(str, Enum):
@@ -36,8 +35,7 @@ class PaymentBaseSchema(BaseModel):
     external_payment_id: Optional[str] = None
     payment_items: list[PaymentItemBaseSchema] = []
 
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class PaymentListSchema(BaseModel):
