+++ src/tests/test_integration/test_profiles.py
@@ -6,7 +6,7 @@ from io import BytesIO
 from PIL import Image
 from sqlalchemy import select, func
 
-from database import UserModel, UserProfileModel
+from database.models.accounts import UserModel, UserProfileModel
 from exceptions import S3FileUploadError
